fix AJ5 lint limits
